second comit

+++ app/Http/Controllers/PostController.php
@@ -4,18 +4,30 @@ namespace App\Http\Controllers;
 
 use App\Models\Post;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class PostController extends Controller
 {
     public function index(){
-        $posts = Post::with(['tags','user'])->latest()->simplePaginate(5);
+        $posts = Post::with(['tags','user'])->latest()->simplePaginate(6);
         return view('posts.index',['posts' => $posts]);
     }
     public function create(){
+        if(Auth::user()->can('create-post'))
         return view('posts.create');
+        else redirect('/login');
     }
     public function show(Post $post){
         return view('posts.show', ['post' => $post]);
+    }
+    public function showUserPosts(){
+        if(Auth::guest()){
+            return redirect('/login');
+        }else{
+            $posts = Auth::user()->posts()->with('user')->latest()->simplePaginate(6);
+        }
+        return view('posts.userPosts',['posts'=>$posts]);
+
     }
     public function store(){
         request()->validate([
@@ -26,16 +38,23 @@ class PostController extends Controller
         Post::create([
             'title' => request('title'),
             'text' => request('text'),
-            'user_id' => 1,
+            'user_id' => Auth::user()->id,
             // 'user_id' => Auth::id(),
         ]);
         return redirect('/posts');
     }
     public function edit(Post $post){
-        return view('posts.edit', ['post' => $post]);
+        if($post->user->isNot(Auth::user()))
+        {
+            return redirect('/posts');
+        }else{
+            return view('posts.edit', ['post' => $post]);
+        }
     }
     public function update(Post $post){
         // authorize
+        if($post->user->isNot(Auth::user()))
+            return redirect('/posts');
         //validate
         request()->validate([
             'title' => ['required'],
@@ -58,6 +77,8 @@ class PostController extends Controller
         return redirect('/posts/'.$post->id);
     }
     public function destroy(Post $post){
+        if($post->user->isNot(Auth::user()))
+            return redirect('/posts');
         $post->delete();
         return redirect('/posts');
     }
